Spit dns decoder into multiple files
